Reorganise
Distributes the stuff in BMX to their relevant modules, refactoring
where necessary to avoid cycles.

Allows use of BMXState and other useful frontend functions / types
from other modules, which will soon be necessary
