+++ src/BMX.hs
@@ -1,6 +1,4 @@
 {-# LANGUAGE NoImplicitPrelude #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE OverloadedStrings #-}
 
 -- | A templating library in the style of <http://handlebarsjs.com Handlebars.js>,
 -- embedded in Haskell for static or server-side rendering.
@@ -60,19 +58,11 @@ module BMX (
   , usingDecorators
   ) where
 
-import           Control.Monad.Identity (Identity)
-import           Data.Map.Strict (Map)
-import qualified Data.Map.Strict as M
-import           Data.Text (Text)
-import           System.IO (IO)
-
-import           BMX.Builtin
+import           BMX.Builtin (defaultState)
 import           BMX.Data
-import           BMX.Function
-import           BMX.Lexer as X (tokenise)
-import           BMX.Parser as X (parse)
-
-import           P
+import           BMX.Debug (debugTemplateIO)
+import           BMX.Eval (renderTemplate, renderTemplateM, renderTemplateIO, partialFromTemplate)
+import           BMX.Parser (templateFromText)
 
 -- $whatsnew
 --
@@ -167,117 +157,3 @@ import           P
 --
 -- See <BMX-Function.html BMX.Function> for details on implementing
 -- custom decorators.
-
--- | BMXState holds the initial rendering environment, i.e. all bound helpers,
--- partials, decorators, and the current variable context.
---
--- The type parameter @m@ refers to the base monad for the registered
--- helpers, partials and decorators. It is commonly 'Identity' or
--- 'IO', though it should be possible to render a 'Template' on top of
--- any monad stack.
-data BMXState m = BMXState
-  { bmxContext :: Context
-  , bmxPartials :: [(Text, Partial m)]
-  , bmxHelpers :: [(Text, Helper m)]
-  , bmxDecorators :: [(Text, Decorator m)]
-  }
-
-instance Monoid (BMXState m) where
-  mempty = BMXState mempty mempty mempty mempty
-  mappend a b = BMXState {
-      bmxContext = bmxContext a <> bmxContext b
-    , bmxPartials = bmxPartials a <> bmxPartials b
-    , bmxHelpers = bmxHelpers a <> bmxHelpers b
-    , bmxDecorators = bmxDecorators a <> bmxDecorators b
-    }
-
--- | The default state: an empty context, all the helpers from
--- 'BMX.Builtin.Helpers.builtinHelpers', and all the decorators from
--- 'BMX.Builtin.Decorators.builtinDecorators'.
-defaultState :: (Applicative m, Monad m) => BMXState m
-defaultState = mempty {
-    bmxHelpers = builtinHelpers
-  , bmxDecorators = builtinDecorators
-  }
-
--- | Set the initial context in an 'BMXState'.
-usingContext :: (Applicative m, Monad m) => BMXState m -> Context -> BMXState m
-usingContext st c = st { bmxContext = c }
-
--- | Add a named collection of templates to the 'BMXState' as partials.
-usingPartials :: (Applicative m, Monad m) => BMXState m -> [(Text, Template)] -> BMXState m
-usingPartials st ps = st { bmxPartials = (fmap . fmap) partialFromTemplate ps <> bmxPartials st }
-
--- | Add a named collection of helpers to the 'BMXState'.
-usingHelpers :: (Applicative m, Monad m) => BMXState m -> [(Text, Helper m)] -> BMXState m
-usingHelpers st hs = st { bmxHelpers = hs <> bmxHelpers st }
-
--- | Add a named collection of decorators to the 'BMXState'.
-usingDecorators :: (Applicative m, Monad m) => BMXState m -> [(Text, Decorator m)] -> BMXState m
-usingDecorators st ds = st { bmxDecorators = ds <> bmxDecorators st }
-
--- | Lex and parse a 'Template' from some 'Text'.
-templateFromText :: Text -> Either BMXError Template
-templateFromText = either convert (first BMXParseError . parse) . tokenise
-  where
-    convert = Left . BMXLexError
-
--- | Apply a 'Template' to some 'BMXState', producing a 'Page'.
---
--- All helpers, partials and decorators must be pure. Use 'renderTemplateIO'
--- if IO helpers are required.
-renderTemplate :: BMXState Identity -> Template -> Either BMXError Page
-renderTemplate bst t = do
-  st <- packState bst
-  first BMXEvalError $ runBMX st (eval t)
-
--- | Apply a 'Template' to a 'BMXState' in some Monad stack, producing
--- a 'Page'.
-renderTemplateM :: (Applicative m, Monad m) => BMXState m -> Template -> m (Either BMXError Page)
-renderTemplateM bst t = either (return . Left) runIt (packState bst)
-  where runIt es = do
-          ep <- runBMXT es (eval t)
-          return (first BMXEvalError ep)
-
--- | Apply a 'Template' to some 'BMXState', producing a 'Page'.
---
--- Helpers, partials and decorators may involve IO. Use @renderTemplate@ if
--- no IO helpers are to be invoked.
-renderTemplateIO :: BMXState IO -> Template -> IO (Either BMXError Page)
-renderTemplateIO bst = renderTemplateM bst
-
-{-# WARNING debugTemplateIO "Do not use 'debugTemplateIO' in production code" #-}
--- | Evaluate a 'Template' against some 'BMXState' with debugging helpers enabled.
---
--- Debugging helpers are enumerated in 'debugHelpers'.
---
--- Debugging helpers may perform IO. Do not use this in production.
-debugTemplateIO :: BMXState IO -> Template -> IO (Either BMXError Page)
-debugTemplateIO st = renderTemplateIO (st `usingHelpers` debugHelpers)
-
--- | Pack the association lists from 'BMXState' into the maps of 'EvalState',
--- throwing errors whenever shadowing is encountered.
-packState :: (Applicative m, Monad m) => BMXState m -> Either BMXError (EvalState m)
-packState bst = do
-  let ctx = [bmxContext bst]
-  partials <- mapUnique (bmxPartials bst)
-  helpers <- mapUnique (bmxHelpers bst)
-  decorators <- mapUnique (bmxDecorators bst)
-  let dta = mempty
-  return EvalState {
-      evalContext = ctx
-    , evalData = dta
-    , evalHelpers = helpers
-    , evalPartials = partials
-    , evalDecorators = decorators
-    }
-
-mapUnique :: [(Text, a)] -> Either BMXError (Map Text a)
-mapUnique = foldM foldFun M.empty
-  where foldFun m (k, v)  = if M.member k m
-          then Left (BMXEvalError (Shadowing k))
-          else Right (M.insert k v m)
-
--- | Produce a 'Partial' from an ordinary 'Template'.
-partialFromTemplate :: (Applicative m, Monad m) => Template -> Partial m
-partialFromTemplate = partial . eval
